// src/types/package.ts
export interface Provider {
  _id: string;
  name: string;
  code: "MTN" | "TELECEL" | "AT" | "AFA" | string;
  description?: string;
  logo?: {
    url: string;
    alt: string;
  };
  isActive: boolean;
  createdBy: string;
  updatedBy?: string;
  isDeleted: boolean;
  deletedAt?: Date;
  deletedBy?: string;
  salesCount: number;
  viewCount: number;
  createdAt: Date;
  updatedAt: Date;
}